+++ noodles-bam/src/record/cigar.rs
@@ -5,26 +5,31 @@ mod ops;
 
 pub use self::{op::Op, ops::Ops};
 
-use std::{fmt, io, ops::Deref};
+use std::{
+    fmt, io,
+    ops::{Deref, DerefMut},
+};
 
 use noodles_sam::{self as sam, record::cigar::op::Kind};
 
 /// BAM record CIGAR.
-pub struct Cigar<'a>(&'a [u32]);
+#[derive(Clone, Default, Eq, PartialEq)]
+pub struct Cigar(Vec<u32>);
 
-impl<'a> Cigar<'a> {
+impl Cigar {
     /// Creates a CIGAR by wrapping raw CIGAR data.
     ///
     /// # Examples
     ///
     /// ```
     /// use noodles_bam::record::Cigar;
-    /// let data = [0x00000240, 0x00000084]; // 36M8S
-    /// let cigar = Cigar::new(&data);
-    /// assert_eq!(*cigar, data);
+    /// let data = vec![0x00000240, 0x00000084]; // 36M8S
+    /// let cigar = Cigar::from(data);
+    /// assert_eq!(**cigar, [0x00000240, 0x00000084]);
     /// ```
-    pub fn new(cigar: &[u32]) -> Cigar<'_> {
-        Cigar(cigar)
+    #[deprecated(since = "0.8.0", note = "Use `Cigar::from::<Vec<u32>>` instead.")]
+    pub fn new(cigar: Vec<u32>) -> Cigar {
+        Self::from(cigar)
     }
 
     /// Returns an iterator over the operations in the CIGAR.
@@ -36,8 +41,8 @@ impl<'a> Cigar<'a> {
     /// use noodles_bam::record::{cigar::Op, Cigar};
     /// use noodles_sam::record::cigar::op::Kind;
     ///
-    /// let data = [0x00000240, 0x00000084]; // 36M8S
-    /// let cigar = Cigar::new(&data);
+    /// let data = vec![0x00000240, 0x00000084]; // 36M8S
+    /// let cigar = Cigar::from(data);
     ///
     /// let mut ops = cigar.ops();
     ///
@@ -47,7 +52,7 @@ impl<'a> Cigar<'a> {
     /// # Ok::<(), Box<dyn std::error::Error>>(())
     /// ```
     pub fn ops(&self) -> Ops<'_> {
-        Ops::new(self.0)
+        Ops::new(&self.0)
     }
 
     /// Calculates the alignment span over the reference sequence.
@@ -63,8 +68,8 @@ impl<'a> Cigar<'a> {
     /// use noodles_bam::record::{cigar::Op, Cigar};
     /// use noodles_sam::record::cigar::op::Kind;
     ///
-    /// let data = [0x00000240, 0x00000043, 0x00000084]; // 36M4D8S
-    /// let cigar = Cigar::new(&data);
+    /// let data = vec![0x00000240, 0x00000043, 0x00000084]; // 36M4D8S
+    /// let cigar = Cigar::from(data);
     ///
     /// assert_eq!(cigar.reference_len()?, 40);
     /// # Ok::<(), io::Error>(())
@@ -99,8 +104,8 @@ impl<'a> Cigar<'a> {
     /// use noodles_bam::record::{cigar::Op, Cigar};
     /// use noodles_sam::record::cigar::op::Kind;
     ///
-    /// let data = [0x00000240, 0x00000043, 0x00000084]; // 36M4D8S
-    /// let cigar = Cigar::new(&data);
+    /// let data = vec![0x00000240, 0x00000043, 0x00000084]; // 36M4D8S
+    /// let cigar = Cigar::from(data);
     ///
     /// assert_eq!(cigar.read_len()?, 44);
     /// # Ok::<(), io::Error>(())
@@ -127,13 +132,13 @@ impl<'a> Cigar<'a> {
     }
 }
 
-impl<'a> fmt::Debug for Cigar<'a> {
+impl<'a> fmt::Debug for Cigar {
     fn fmt(&self, fmt: &mut fmt::Formatter<'_>) -> fmt::Result {
         fmt.debug_list().entries(self.ops()).finish()
     }
 }
 
-impl<'a> fmt::Display for Cigar<'a> {
+impl<'a> fmt::Display for Cigar {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         for result in self.ops() {
             let op = result.map_err(|_| fmt::Error)?;
@@ -144,18 +149,30 @@ impl<'a> fmt::Display for Cigar<'a> {
     }
 }
 
-impl<'a> Deref for Cigar<'a> {
-    type Target = [u32];
+impl Deref for Cigar {
+    type Target = Vec<u32>;
 
     fn deref(&self) -> &Self::Target {
-        self.0
+        &self.0
     }
 }
 
-impl<'a> TryFrom<Cigar<'a>> for sam::record::Cigar {
+impl DerefMut for Cigar {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        &mut self.0
+    }
+}
+
+impl From<Vec<u32>> for Cigar {
+    fn from(cigar: Vec<u32>) -> Self {
+        Self(cigar)
+    }
+}
+
+impl TryFrom<&Cigar> for sam::record::Cigar {
     type Error = io::Error;
 
-    fn try_from(cigar: Cigar<'_>) -> Result<Self, Self::Error> {
+    fn try_from(cigar: &Cigar) -> Result<Self, Self::Error> {
         let mut ops = Vec::new();
 
         for result in cigar.ops() {
@@ -173,8 +190,8 @@ mod tests {
 
     #[test]
     fn test_from_bytes() -> Result<(), Box<dyn std::error::Error>> {
-        let data = [0x00000240, 0x00000362]; // 36MD54
-        let cigar = Cigar::new(&data);
+        let data = vec![0x00000240, 0x00000362]; // 36MD54
+        let cigar = Cigar::from(data);
 
         let mut ops = cigar.ops();
 
@@ -189,10 +206,10 @@ mod tests {
     fn test_try_from_cigar_for_sam_record_cigar() -> io::Result<()> {
         use sam::record::cigar::{op, Op};
 
-        let data = [0x00000240, 0x00000362]; // 36MD54
-        let cigar = Cigar::new(&data);
+        let data = vec![0x00000240, 0x00000362]; // 36MD54
+        let cigar = Cigar::from(data);
 
-        let actual = sam::record::Cigar::try_from(cigar)?;
+        let actual = sam::record::Cigar::try_from(&cigar)?;
         let expected = sam::record::Cigar::from(vec![
             Op::new(op::Kind::Match, 36),
             Op::new(op::Kind::Deletion, 54),